refactor(programs): fold MigrationReport back in, wrap failures in _migrate_group
Supersedes the standalone migration_report module: keeping the value
object in migrate.py avoids the averaging artefact that made the OO
ratchet refuse the split, while the crash-safe loop drops the old
entries-generator so the F3 try/except adds no net complexity. Condensed
the module and method docstrings (the per-method rationale now lives once
in the module docstring), improving module_size.
